requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
python-dotenv>=1.0.0